requirements.txt
numpy
pillow
tqdm
matplotlib
tensorflow<2.0.0
tensorflowjs==0.8.5
emnist